Added more gui stuff. Fruit icons.

+++ VerticalLayout.cpp
@@ -8,7 +8,7 @@ VerticalLayout::VerticalLayout(GuiElement *parent) : Layout(parent)
 
 void VerticalLayout::drawChildren()
 {
-	for (int i=0; i<m_children->size(); i++)
+	for (int i=0; i < (int)m_children->size(); i++)
 	{
 		GuiElement *child = (*m_children)[i];
 		child->draw();
@@ -20,7 +20,7 @@ void VerticalLayout::drawChildren()
 void VerticalLayout::onClick(int nButton, int nState, float nX, float nY)
 {
 	float nCurYPos = 0;
-	for (int i=0; i<m_children->size(); i++)
+	for (int i=0; i < (int)m_children->size(); i++)
 	{
 		GuiElement *child = (*m_children)[i];
 		float nChildClickX = (nX - child->getXPos()) / child->getWidth();
